Move Proxy.initialize within FlagParser.initialize.  Also move other staticmethods from within proxy class into utils
