+++ examples/tcp_echo_server.py
@@ -11,7 +11,7 @@
 import time
 from typing import Optional
 
-from proxy.proxy import Proxy
+from proxy.common.flag import FlagParser
 from proxy.core.acceptor import AcceptorPool
 from proxy.core.base import BaseTcpServerHandler
 
@@ -31,7 +31,8 @@ class EchoServerHandler(BaseTcpServerHandler):
 def main() -> None:
     # This example requires `threadless=True`
     with AcceptorPool(
-        flags=Proxy.initialize(port=12345, num_workers=1, threadless=True),
+        flags=FlagParser.initialize(
+            port=12345, num_workers=1, threadless=True),
         work_klass=EchoServerHandler,
     ):
         try:
