+++ tests/http/test_protocol_handler.py
@@ -15,9 +15,9 @@ import base64
 from typing import cast
 from unittest import mock
 
-from proxy.proxy import Proxy
+from proxy.common.flag import FlagParser
 from proxy.common.version import __version__
-from proxy.common.utils import bytes_
+from proxy.common.utils import bytes_, load_plugins
 from proxy.common.constants import CRLF, PLUGIN_HTTP_PROXY, PLUGIN_PROXY_AUTH, PLUGIN_WEB_SERVER
 from proxy.core.connection import TcpClientConnection
 from proxy.http.parser import HttpParser
@@ -41,8 +41,8 @@ class TestHttpProtocolHandler(unittest.TestCase):
         self._conn = mock_fromfd.return_value
 
         self.http_server_port = 65535
-        self.flags = Proxy.initialize()
-        self.flags.plugins = Proxy.load_plugins([
+        self.flags = FlagParser.initialize()
+        self.flags.plugins = load_plugins([
             bytes_(PLUGIN_HTTP_PROXY),
             bytes_(PLUGIN_WEB_SERVER),
         ])
@@ -212,10 +212,10 @@ class TestHttpProtocolHandler(unittest.TestCase):
     ) -> None:
         self._conn = mock_fromfd.return_value
         self.mock_selector_for_client_read(mock_selector)
-        flags = Proxy.initialize(
+        flags = FlagParser.initialize(
             auth_code=base64.b64encode(b'user:pass'),
         )
-        flags.plugins = Proxy.load_plugins([
+        flags.plugins = load_plugins([
             bytes_(PLUGIN_HTTP_PROXY),
             bytes_(PLUGIN_WEB_SERVER),
             bytes_(PLUGIN_PROXY_AUTH),
@@ -250,10 +250,10 @@ class TestHttpProtocolHandler(unittest.TestCase):
         server.connect.return_value = True
         server.buffer_size.return_value = 0
 
-        flags = Proxy.initialize(
+        flags = FlagParser.initialize(
             auth_code=base64.b64encode(b'user:pass'),
         )
-        flags.plugins = Proxy.load_plugins([
+        flags.plugins = load_plugins([
             bytes_(PLUGIN_HTTP_PROXY),
             bytes_(PLUGIN_WEB_SERVER),
         ])
@@ -305,10 +305,10 @@ class TestHttpProtocolHandler(unittest.TestCase):
             mock_selector, server,
         )
 
-        flags = Proxy.initialize(
+        flags = FlagParser.initialize(
             auth_code=base64.b64encode(b'user:pass'),
         )
-        flags.plugins = Proxy.load_plugins([
+        flags.plugins = load_plugins([
             bytes_(PLUGIN_HTTP_PROXY),
             bytes_(PLUGIN_WEB_SERVER),
         ])
